requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.6
Flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonify==0.5
MarkupSafe==3.0.3
oauthlib==3.3.1
pycparser==3.0
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.7
